requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
lxml==5.2.2
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pyaes==1.6.1
pydantic==2.7.4
pydantic_core==2.18.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==13.4
pytz==2024.1
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
telegram==0.0.1
tornado==6.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
